Make the expansion easier to test (#11)
In previous implementation, there was a `dep_doc` (resp. `dev_dep_doc`
macro, which expanded to the concatenation of the following:
  - the beginning of a code snippet with the section `[dependencies]`
    (resp. `[dev-dependencies])`,
  - a call to package_import with the correct package name and version
    passed as argument,
  - the ned of the code snippet.

The problem with this approach is that it was not possible to test the
complete expansion of a macro, given a crate name, a version and some
other information. This even resulted to a bug where we forgot to pass
the so called "other information" to the package_import. See #10 for
more.

This commit allows for simple testing by adding an additional
intermediate step. The `dep_doc` (resp. `dev_dep_doc`) macro now expands
to a call to `dep_doc_inner` (resp. `dev_dep_doc_inner`), with the crate
name and version guessed from the environment and the
"other information". `dep_doc_inner` (resp. `dev_dep_doc`) is then
responsible for generating the concatenation described previously.

As such, most of the expansion process can be tested by calling
`dep_doc_inner` (resp. `dev_dep_doc_inner`).

A bunch of tests have been added in the process in order to ensure that
the previously linked bug won't happen again.